Rename path component `rfc4122` to `standard`
