+++ src/main/java/com/github/f4b6a3/uuid/factory/standard/DceSecurityFactory.java
@@ -22,7 +22,7 @@
  * SOFTWARE.
  */
 
-package com.github.f4b6a3.uuid.factory.rfc4122;
+package com.github.f4b6a3.uuid.factory.standard;
 
 import java.util.UUID;
 import java.util.concurrent.atomic.AtomicInteger;
