+++ src/main/java/com/github/f4b6a3/uuid/factory/standard/TimeOrderedEpochFactory.java
@@ -22,7 +22,7 @@
  * SOFTWARE.
  */
 
-package com.github.f4b6a3.uuid.factory.rfc4122;
+package com.github.f4b6a3.uuid.factory.standard;
 
 import java.time.Clock;
 import java.util.Random;
