+++ src/test/java/com/github/f4b6a3/uuid/TestSuite.java
@@ -25,13 +25,13 @@ import com.github.f4b6a3.uuid.factory.nonstandard.PrefixCombFactoryTest;
 import com.github.f4b6a3.uuid.factory.nonstandard.ShortPrefixCombFactoryTest;
 import com.github.f4b6a3.uuid.factory.nonstandard.ShortSuffixCombFactoryTest;
 import com.github.f4b6a3.uuid.factory.nonstandard.SuffixCombFactoryTest;
-import com.github.f4b6a3.uuid.factory.rfc4122.DceSecurityFactoryTest;
-import com.github.f4b6a3.uuid.factory.rfc4122.NameBasedMd5FactoryTest;
-import com.github.f4b6a3.uuid.factory.rfc4122.NameBasedSha1FactoryTest;
-import com.github.f4b6a3.uuid.factory.rfc4122.RandomBasedFactoryTest;
-import com.github.f4b6a3.uuid.factory.rfc4122.TimeBasedFactoryTest;
-import com.github.f4b6a3.uuid.factory.rfc4122.TimeOrderedEpochFactoryTest;
-import com.github.f4b6a3.uuid.factory.rfc4122.TimeOrderedFactoryTest;
+import com.github.f4b6a3.uuid.factory.standard.DceSecurityFactoryTest;
+import com.github.f4b6a3.uuid.factory.standard.NameBasedMd5FactoryTest;
+import com.github.f4b6a3.uuid.factory.standard.NameBasedSha1FactoryTest;
+import com.github.f4b6a3.uuid.factory.standard.RandomBasedFactoryTest;
+import com.github.f4b6a3.uuid.factory.standard.TimeBasedFactoryTest;
+import com.github.f4b6a3.uuid.factory.standard.TimeOrderedEpochFactoryTest;
+import com.github.f4b6a3.uuid.factory.standard.TimeOrderedFactoryTest;
 import com.github.f4b6a3.uuid.alt.GUIDTest;
 import com.github.f4b6a3.uuid.codec.BinaryCodecTest;
 import com.github.f4b6a3.uuid.codec.StringCodecTest;
