+++ src/test/java/com/github/f4b6a3/uuid/factory/function/impl/DefaultClockSeqFunctionTest.java
@@ -4,7 +4,7 @@ import org.junit.Before;
 import org.junit.Test;
 
 import com.github.f4b6a3.uuid.factory.function.ClockSeqFunction;
-import com.github.f4b6a3.uuid.factory.rfc4122.TimeBasedFactory;
+import com.github.f4b6a3.uuid.factory.standard.TimeBasedFactory;
 import com.github.f4b6a3.uuid.util.UuidTime;
 import com.github.f4b6a3.uuid.util.UuidUtil;
